refactor: sessionLog 패키지를 session > log 패키지로 변경 및 이동
