+++ src/main/java/koza/licensemanagementservice/domain/session/log/controller/SessionLogController.java
@@ -1,12 +1,12 @@
-package koza.licensemanagementservice.domain.sessionLog.controller;
+package koza.licensemanagementservice.domain.session.log.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import koza.licensemanagementservice.domain.session.log.dto.DailyUsageResponse;
+import koza.licensemanagementservice.domain.session.log.dto.SessionHistoryResponse;
+import koza.licensemanagementservice.domain.session.log.service.SessionLogService;
 import koza.licensemanagementservice.global.common.ApiResponse;
 import koza.licensemanagementservice.auth.dto.CustomUser;
-import koza.licensemanagementservice.domain.sessionLog.service.SessionLogService;
-import koza.licensemanagementservice.domain.sessionLog.dto.DailyUsageResponse;
-import koza.licensemanagementservice.domain.sessionLog.dto.SessionHistoryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
