+++ src/main/java/koza/licensemanagementservice/domain/session/log/repository/SessionLogRepositoryCustomImpl.java
@@ -1,11 +1,11 @@
-package koza.licensemanagementservice.domain.sessionLog.repository;
+package koza.licensemanagementservice.domain.session.log.repository;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.PathBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import koza.licensemanagementservice.domain.sessionLog.entity.SessionLog;
+import koza.licensemanagementservice.domain.session.log.entity.SessionLog;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -17,7 +17,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static koza.licensemanagementservice.domain.license.entity.QLicense.license;
-import static koza.licensemanagementservice.domain.sessionLog.entity.QSessionLog.sessionLog;
+import static koza.licensemanagementservice.domain.session.log.entity.QSessionLog.sessionLog;
 
 
 @RequiredArgsConstructor
